Update for compatibility with flask v2.3.2
In version 2.3.0 of flask, the JSONEncoder class was removed and the
app.json_encoder attribute was removed. The result that a run of
the tests using flask==2.3.0 would fail. This change updates the
use of flask's json mechanisms to be in line with the new standard,
and similarly updates the tests to utilize those mechanisms correctly.